Remove root of repo from path
This removes all imports that used the fact that the root of the repo
was automatically added to the list of paths to import from. This makes it
far easier for other tools to be used without much change because most
tools don't implicitly add the root folder to the path.